+++ lib/ansible/module_utils/network/cnos/cnos.py
@@ -1363,523 +1363,6 @@ def portChannelConfig(module, prompt, answer):
 # EOM
 
 
-def vlanConfig(module, prompt, answer):
-
-    retVal = ''
-    # Wait time to get response from server
-    vlanArg1 = module.params['vlanArg1']
-    vlanArg2 = module.params['vlanArg2']
-    vlanArg3 = module.params['vlanArg3']
-    vlanArg4 = module.params['vlanArg4']
-    vlanArg5 = module.params['vlanArg5']
-    deviceType = module.params['deviceType']
-    # vlan config command happens here.
-    command = 'vlan '
-
-    if(vlanArg1 == "access-map"):
-        # debugOutput("access-map ")
-        command = command + vlanArg1 + ' '
-        value = checkSanityofVariable(
-            deviceType, "vlan_access_map_name", vlanArg2)
-        if(value == "ok"):
-            command = command + vlanArg2
-            # debugOutput(command)
-            cmd = [{'command': command, 'prompt': None, 'answer': None}]
-            retVal = retVal + vlanAccessMapConfig(module, cmd)
-            return retVal
-        else:
-            retVal = "Error-130"
-            return retVal
-
-    elif(vlanArg1 == "dot1q"):
-        # debugOutput("dot1q")
-        command = command + vlanArg1 + " tag native "
-        if(vlanArg2 is not None):
-            value = checkSanityofVariable(
-                deviceType, "vlan_dot1q_tag", vlanArg2)
-            if(value == "ok"):
-                command = command + vlanArg2
-            else:
-                retVal = "Error-131"
-                return retVal
-
-    elif(vlanArg1 == "filter"):
-        # debugOutput( "filter")
-        command = command + vlanArg1 + " "
-        if(vlanArg2 is not None):
-            value = checkSanityofVariable(
-                deviceType, "vlan_filter_name", vlanArg2)
-            if(value == "ok"):
-                command = command + vlanArg2 + " vlan-list "
-                value = checkSanityofVariable(deviceType, "vlan_id", vlanArg3)
-                if(value == "ok"):
-                    command = command + vlanArg3
-                else:
-                    value = checkSanityofVariable(
-                        deviceType, "vlan_id_range", vlanArg3)
-                    if(value == "ok"):
-                        command = command + vlanArg3
-                    else:
-                        retVal = "Error-133"
-                    return retVal
-            else:
-                retVal = "Error-132"
-                return retVal
-
-    else:
-        value = checkSanityofVariable(deviceType, "vlan_id", vlanArg1)
-        if(value == "ok"):
-            retVal = createVlan(module, '(config-vlan)#', None)
-            return retVal
-        else:
-            value = checkSanityofVariable(
-                deviceType, "vlan_id_range", vlanArg1)
-            if(value == "ok"):
-                retVal = createVlan(module, '(config-vlan)#', None)
-                return retVal
-            retVal = "Error-133"
-            return retVal
-
-    # debugOutput(command)
-    cmd = [{'command': command, 'prompt': None, 'answer': None}]
-    retVal = retVal + str(run_cnos_commands(module, cmd))
-    return retVal
-# EOM
-
-
-def vlanAccessMapConfig(module, cmd):
-    retVal = ''
-    # Wait time to get response from server
-    command = ''
-    vlanArg3 = module.params['vlanArg3']
-    vlanArg4 = module.params['vlanArg4']
-    vlanArg5 = module.params['vlanArg5']
-    deviceType = module.params['deviceType']
-    if(vlanArg3 == "action"):
-        command = command + vlanArg3 + ' '
-        value = checkSanityofVariable(
-            deviceType, "vlan_accessmap_action", vlanArg4)
-        if(value == "ok"):
-            command = command + vlanArg4
-        else:
-            retVal = "Error-135"
-            return retVal
-    elif(vlanArg3 == "match"):
-        command = command + vlanArg3 + ' '
-        if(vlanArg4 == "ip" or vlanArg4 == "mac"):
-            command = command + vlanArg4 + ' address '
-            value = checkSanityofVariable(
-                deviceType, "vlan_access_map_name", vlanArg5)
-            if(value == "ok"):
-                command = command + vlanArg5
-            else:
-                retVal = "Error-136"
-                return retVal
-        else:
-            retVal = "Error-137"
-            return retVal
-    elif(vlanArg3 == "statistics"):
-        command = vlanArg3 + " per-entry"
-    else:
-        retVal = "Error-138"
-        return retVal
-
-    inner_cmd = [{'command': command, 'prompt': None, 'answer': None}]
-    cmd.extend(inner_cmd)
-    retVal = retVal + str(run_cnos_commands(module, cmd))
-    # debugOutput(command)
-    return retVal
-# EOM
-
-
-def checkVlanNameNotAssigned(module, prompt, answer):
-    retVal = "ok"
-    vlanId = module.params['vlanArg1']
-    vlanName = module.params['vlanArg3']
-    command = "show vlan id " + vlanId
-    cmd = [{'command': command, 'prompt': None, 'answer': None}]
-    retVal = str(run_cnos_commands(module, cmd))
-    if(retVal.find('Error') != -1):
-        command = "display vlan id " + vlanId
-        retVal = str(run_cnos_commands(module, cmd))
-    if(retVal.find(vlanName) != -1):
-        return "Nok"
-    else:
-        return "ok"
-# EOM
-
-
-# Utility Method to create vlan
-def createVlan(module, prompt, answer):
-
-    # vlan config command happens here. It creates if not present
-    vlanArg1 = module.params['vlanArg1']
-    vlanArg2 = module.params['vlanArg2']
-    vlanArg3 = module.params['vlanArg3']
-    vlanArg4 = module.params['vlanArg4']
-    vlanArg5 = module.params['vlanArg5']
-    deviceType = module.params['deviceType']
-    retVal = ''
-    command = 'vlan ' + vlanArg1
-    # debugOutput(command)
-    cmd = [{'command': command, 'prompt': None, 'answer': None}]
-    command = ""
-    if(vlanArg2 == "name"):
-        # debugOutput("name")
-        command = vlanArg2 + " "
-        value = checkSanityofVariable(deviceType, "vlan_name", vlanArg3)
-        if(value == "ok"):
-            value = checkVlanNameNotAssigned(module, prompt, answer)
-            if(value == "ok"):
-                command = command + vlanArg3
-            else:
-                retVal = retVal + 'VLAN Name is already assigned \n'
-                command = "\n"
-        else:
-            retVal = "Error-139"
-            return retVal
-    elif (vlanArg2 == "flood"):
-        # debugOutput("flood")
-        command = vlanArg2 + " "
-        value = checkSanityofVariable(deviceType, "vlan_flood", vlanArg3)
-        if(value == "ok"):
-            command = command + vlanArg3
-        else:
-            retVal = "Error-140"
-            return retVal
-
-    elif(vlanArg2 == "state"):
-        # debugOutput("state")
-        command = vlanArg2 + " "
-        value = checkSanityofVariable(deviceType, "vlan_state", vlanArg3)
-        if(value == "ok"):
-            command = command + vlanArg3
-        else:
-            retVal = "Error-141"
-            return retVal
-
-    elif(vlanArg2 == "ip"):
-        # debugOutput("ip")
-        command = vlanArg2 + " igmp snooping "
-        # debugOutput("vlanArg3")
-        if(vlanArg3 is None or vlanArg3 == ""):
-            # debugOutput("None or empty")
-            command = command.strip()
-        elif(vlanArg3 == "fast-leave"):
-            # debugOutput("fast-leave")
-            command = command + vlanArg3
-
-        elif (vlanArg3 == "last-member-query-interval"):
-            # debugOutput("last-member-query-interval")
-            command = command + vlanArg3 + " "
-            value = checkSanityofVariable(
-                deviceType, "vlan_last_member_query_interval", vlanArg4)
-            if(value == "ok"):
-                command = command + vlanArg4
-            else:
-                retVal = "Error-142"
-                return retVal
-
-        elif (vlanArg3 == "querier"):
-            # debugOutput("querier")
-            command = command + vlanArg3 + " "
-            value = checkSanityofVariable(deviceType, "vlan_querier", vlanArg4)
-            if(value == "ok"):
-                command = command + vlanArg4
-            else:
-                retVal = "Error-143"
-                return retVal
-        elif (vlanArg3 == "querier-timeout"):
-            # debugOutput("querier-timeout")
-            command = command + vlanArg3 + " "
-            value = checkSanityofVariable(
-                deviceType, "vlan_querier_timeout", vlanArg4)
-            if(value == "ok"):
-                command = command + vlanArg4
-            else:
-                retVal = "Error-144"
-                return retVal
-        elif (vlanArg3 == "query-interval"):
-            # debugOutput("query-interval")
-            command = command + vlanArg3 + " "
-            value = checkSanityofVariable(
-                deviceType, "vlan_query_interval", vlanArg4)
-            if(value == "ok"):
-                command = command + vlanArg4
-            else:
-                retVal = "Error-145"
-                return retVal
-        elif (vlanArg3 == "query-max-response-time"):
-            # debugOutput("query-max-response-time")
-            command = command + vlanArg3 + " "
-            value = checkSanityofVariable(
-                deviceType, "vlan_query_max_response_time", vlanArg4)
-            if(value == "ok"):
-                command = command + vlanArg4
-            else:
-                retVal = "Error-146"
-                return retVal
-        elif (vlanArg3 == "report-suppression"):
-            # debugOutput("report-suppression")
-            command = command + vlanArg3
-
-        elif (vlanArg3 == "robustness-variable"):
-            # debugOutput("robustness-variable")
-            command = command + vlanArg3 + " "
-            value = checkSanityofVariable(
-                deviceType, "vlan_robustness_variable", vlanArg4)
-            if(value == "ok"):
-                command = command + vlanArg4
-            else:
-                retVal = "Error-147"
-                return retVal
-        elif (vlanArg3 == "startup-query-count"):
-            # debugOutput("startup-query-count")
-            command = command + vlanArg3 + " "
-            value = checkSanityofVariable(
-                deviceType, "vlan_startup_query_count", vlanArg4)
-            if(value == "ok"):
-                command = command + vlanArg4
-            else:
-                retVal = "Error-148"
-                return retVal
-        elif (vlanArg3 == "startup-query-interval"):
-            # debugOutput("startup-query-interval")
-            command = command + vlanArg3 + " "
-            value = checkSanityofVariable(
-                deviceType, "vlan_startup_query_interval", vlanArg4)
-            if(value == "ok"):
-                command = command + vlanArg4
-            else:
-                retVal = "Error-149"
-                return retVal
-        elif (vlanArg3 == "static-group"):
-            retVal = "Error-102"
-            return retVal
-        elif (vlanArg3 == "version"):
-            # debugOutput("version")
-            command = command + vlanArg3 + " "
-            value = checkSanityofVariable(
-                deviceType, "vlan_snooping_version", vlanArg4)
-            if(value == "ok"):
-                command = command + vlanArg4
-            else:
-                retVal = "Error-150"
-                return retVal
-        elif (vlanArg3 == "mrouter"):
-            # debugOutput("mrouter")
-            command = command + vlanArg3 + " interface "
-            if(vlanArg4 == "ethernet"):
-                command = command + vlanArg4 + " "
-                value = checkSanityofVariable(
-                    deviceType, "vlan_ethernet_interface", vlanArg5)
-                if(value == "ok"):
-                    command = command + vlanArg5
-                else:
-                    retVal = "Error-151"
-                    return retVal
-            elif(vlanArg4 == "port-aggregation"):
-                command = command + vlanArg4 + " "
-                value = checkSanityofVariable(
-                    deviceType, "vlan_portagg_number", vlanArg5)
-                if(value == "ok"):
-                    command = command + vlanArg5
-                else:
-                    retVal = "Error-152"
-                    return retVal
-            else:
-                retVal = "Error-153"
-                return retVal
-        else:
-            command = command + vlanArg3
-
-    else:
-        retVal = "Error-154"
-        return retVal
-    inner_cmd = [{'command': command, 'prompt': None, 'answer': None}]
-    cmd.extend(inner_cmd)
-    retVal = retVal + str(run_cnos_commands(module, cmd))
-    # debugOutput(command)
-    return retVal
-# EOM
-
-
-def vlagConfig(module, prompt, answer):
-
-    retVal = ''
-    # vlag config command happens here.
-    command = 'vlag '
-
-    vlagArg1 = module.params['vlagArg1']
-    vlagArg2 = module.params['vlagArg2']
-    vlagArg3 = module.params['vlagArg3']
-    vlagArg4 = module.params['vlagArg4']
-    deviceType = module.params['deviceType']
-
-    if(vlagArg1 == "enable"):
-        # debugOutput("enable")
-        command = command + vlagArg1 + " "
-
-    elif(vlagArg1 == "auto-recovery"):
-        # debugOutput("auto-recovery")
-        command = command + vlagArg1 + " "
-        value = checkSanityofVariable(
-            deviceType, "vlag_auto_recovery", vlagArg2)
-        if(value == "ok"):
-            command = command + vlagArg2
-        else:
-            retVal = "Error-160"
-            return retVal
-
-    elif(vlagArg1 == "config-consistency"):
-        # debugOutput("config-consistency")
-        command = command + vlagArg1 + " "
-        value = checkSanityofVariable(
-            deviceType, "vlag_config_consistency", vlagArg2)
-        if(value == "ok"):
-            command = command + vlagArg2
-        else:
-            retVal = "Error-161"
-            return retVal
-
-    elif(vlagArg1 == "isl"):
-        # debugOutput("isl")
-        command = command + vlagArg1 + " port-channel "
-        value = checkSanityofVariable(
-            deviceType, "vlag_port_aggregation", vlagArg2)
-        if(value == "ok"):
-            command = command + vlagArg2
-        else:
-            retVal = "Error-162"
-            return retVal
-
-    elif(vlagArg1 == "mac-address-table"):
-        # debugOutput("mac-address-table")
-        command = command + vlagArg1 + " refresh"
-
-    elif(vlagArg1 == "peer-gateway"):
-        # debugOutput("peer-gateway")
-        command = command + vlagArg1 + " "
-
-    elif(vlagArg1 == "priority"):
-        # debugOutput("priority")
-        command = command + vlagArg1 + " "
-        value = checkSanityofVariable(deviceType, "vlag_priority", vlagArg2)
-        if(value == "ok"):
-            command = command + vlagArg2
-        else:
-            retVal = "Error-163"
-            return retVal
-
-    elif(vlagArg1 == "startup-delay"):
-        # debugOutput("startup-delay")
-        command = command + vlagArg1 + " "
-        value = checkSanityofVariable(
-            deviceType, "vlag_startup_delay", vlagArg2)
-        if(value == "ok"):
-            command = command + vlagArg2
-        else:
-            retVal = "Error-164"
-            return retVal
-
-    elif(vlagArg1 == "tier-id"):
-        # debugOutput("tier-id")
-        command = command + vlagArg1 + " "
-        value = checkSanityofVariable(deviceType, "vlag_tier_id", vlagArg2)
-        if(value == "ok"):
-            command = command + vlagArg2
-        else:
-            retVal = "Error-165"
-            return retVal
-
-    elif(vlagArg1 == "vrrp"):
-        # debugOutput("vrrp")
-        command = command + vlagArg1 + " active"
-
-    elif(vlagArg1 == "instance"):
-        # debugOutput("instance")
-        command = command + vlagArg1 + " "
-        value = checkSanityofVariable(deviceType, "vlag_instance", vlagArg2)
-        if(value == "ok"):
-            command = command + vlagArg2
-            if(vlagArg3 is not None):
-                command = command + " port-channel "
-                value = checkSanityofVariable(
-                    deviceType, "vlag_port_aggregation", vlagArg3)
-                if(value == "ok"):
-                    command = command + vlagArg3
-                else:
-                    retVal = "Error-162"
-                    return retVal
-            else:
-                command = command + " enable "
-        else:
-            retVal = "Error-166"
-            return retVal
-
-    elif(vlagArg1 == "hlthchk"):
-        # debugOutput("hlthchk")
-        command = command + vlagArg1 + " "
-        value = checkSanityofVariable(
-            deviceType, "vlag_hlthchk_options", vlagArg2)
-        if(value == "ok"):
-            if(vlagArg2 == "keepalive-attempts"):
-                value = checkSanityofVariable(
-                    deviceType, "vlag_keepalive_attempts", vlagArg3)
-                if(value == "ok"):
-                    command = command + vlagArg2 + " " + vlagArg3
-                else:
-                    retVal = "Error-167"
-                    return retVal
-            elif(vlagArg2 == "keepalive-interval"):
-                value = checkSanityofVariable(
-                    deviceType, "vlag_keepalive_interval", vlagArg3)
-                if(value == "ok"):
-                    command = command + vlagArg2 + " " + vlagArg3
-                else:
-                    retVal = "Error-168"
-                    return retVal
-            elif(vlagArg2 == "retry-interval"):
-                value = checkSanityofVariable(
-                    deviceType, "vlag_retry_interval", vlagArg3)
-                if(value == "ok"):
-                    command = command + vlagArg2 + " " + vlagArg3
-                else:
-                    retVal = "Error-169"
-                    return retVal
-            elif(vlagArg2 == "peer-ip"):
-                # Here I am not taking care of IPV6 option.
-                value = checkSanityofVariable(
-                    deviceType, "vlag_peerip", vlagArg3)
-                if(value == "ok"):
-                    command = command + vlagArg2 + " " + vlagArg3
-                    if(vlagArg4 is not None):
-                        value = checkSanityofVariable(
-                            deviceType, "vlag_peerip_vrf", vlagArg4)
-                        if(value == "ok"):
-                            command = command + " vrf " + vlagArg4
-                        else:
-                            retVal = "Error-170"
-                            return retVal
-        else:
-            retVal = "Error-171"
-            return retVal
-
-    else:
-        retVal = "Error-172"
-        return retVal
-
-    # debugOutput(command)
-    cmd = [{'command': command, 'prompt': None, 'answer': None}]
-    retVal = retVal + str(run_cnos_commands(module, cmd))
-    return retVal
-# EOM
-
-# Utility Method to back up the start up config
-# This method supports only TFTP or FTP
-# Tuning of timeout parameter is pending
-
-
 def doStartupConfigBackUp(
         protocol, timeout, confServerIp, confPath, confServerUser,
         confServerPwd, obj):
