+++ lib/ansible/modules/network/cnos/cnos_vlag.py
@@ -33,7 +33,8 @@ DOCUMENTATION = '''
 ---
 module: cnos_vlag
 author: "Anil Kumar Muraleedharan (@amuraleedhar)"
-short_description: Manage VLAG resources and attributes on devices running Lenovo CNOS
+short_description: Manage VLAG resources and attributes on devices running
+ Lenovo CNOS
 description:
     - This module allows you to work with virtual Link Aggregation Groups
      (vLAG) related configurations. The operators used are overloaded to ensure
@@ -44,34 +45,44 @@ description:
      these arguments, see [Overloaded Variables].
      This module uses SSH to manage network device configuration.
      The results of the operation will be placed in a directory named 'results'
-     that must be created by the user in their local directory to where the playbook is run.
-     For more information about this module from Lenovo and customizing it usage for your
-     use cases, please visit U(http://systemx.lenovofiles.com/help/index.jsp?topic=%2Fcom.lenovo.switchmgt.ansible.doc%2Fcnos_vlag.html)
+     that must be created by the user in their local directory to where the
+     playbook is run. For more information about this module from Lenovo and
+     customizing it usage for your use cases, please visit
+     U(http://systemx.lenovofiles.com/help/index.jsp?topic=%2Fcom.lenovo.switchmgt.ansible.doc%2Fcnos_vlag.html)
 version_added: "2.3"
 extends_documentation_fragment: cnos
 options:
   vlagArg1:
     description:
-      - This is an overloaded vlag first argument. Usage of this argument can be found is the User Guide referenced above.
+      - This is an overloaded vlag first argument. Usage of this argument can
+        be found is the User Guide referenced above.
     required: Yes
     default: Null
-    choices: [enable, auto-recovery,config-consistency,isl,mac-address-table,peer-gateway,priority,startup-delay,tier-id,vrrp,instance,hlthchk]
+    choices: [enable, auto-recovery,config-consistency,isl,mac-address-table,
+              peer-gateway,priority,startup-delay,tier-id,vrrp,instance,hlthchk]
   vlagArg2:
     description:
-      - This is an overloaded vlag second argument. Usage of this argument can be found is the User Guide referenced above.
+      - This is an overloaded vlag second argument. Usage of this argument can
+        be found is the User Guide referenced above.
     required: No
     default: Null
-    choices: [Interval in seconds,disable or strict,Port Aggregation Number,VLAG priority,Delay time in seconds,VLAG tier-id value,
-        VLAG instance number,keepalive-attempts,keepalive-interval,retry-interval,peer-ip]
+    choices: [Interval in seconds,disable or strict,Port Aggregation Number,
+              VLAG priority,Delay time in seconds,VLAG tier-id value,
+              VLAG instance number,keepalive-attempts,keepalive-interval,
+              retry-interval,peer-ip]
   vlagArg3:
     description:
-      - This is an overloaded vlag third argument. Usage of this argument can be found is the User Guide referenced above.
+      - This is an overloaded vlag third argument. Usage of this argument can
+        be found is the User Guide referenced above.
     required: No
     default: Null
-    choices: [enable or port-aggregation,Number of keepalive attempts,Interval in seconds,Interval in seconds,VLAG health check peer IP4 address]
+    choices: [enable or port-aggregation,Number of keepalive attempts,
+              Interval in seconds,Interval in seconds,
+              VLAG health check peer IP4 address]
   vlagArg4:
     description:
-      - This is an overloaded vlag fourth argument. Usage of this argument can be found is the User Guide referenced above.
+      - This is an overloaded vlag fourth argument. Usage of this argument can
+        be found is the User Guide referenced above.
     required: No
     default: Null
     choices: [Port Aggregation Number,default or management]
@@ -79,7 +90,8 @@ options:
 '''
 EXAMPLES = '''
 
-Tasks : The following are examples of using the module cnos_vlag. These are written in the main.yml file of the tasks directory.
+Tasks : The following are examples of using the module cnos_vlag. These are
+        written in the main.yml file of the tasks directory.
 ---
 - name: Test Vlag  - enable
   cnos_vlag:
@@ -268,6 +280,180 @@ from ansible.module_utils.basic import AnsibleModule
 from collections import defaultdict
 
 
+def vlagConfig(module, prompt, answer):
+
+    retVal = ''
+    # vlag config command happens here.
+    command = 'vlag '
+
+    vlagArg1 = module.params['vlagArg1']
+    vlagArg2 = module.params['vlagArg2']
+    vlagArg3 = module.params['vlagArg3']
+    vlagArg4 = module.params['vlagArg4']
+    deviceType = module.params['deviceType']
+
+    if(vlagArg1 == "enable"):
+        # debugOutput("enable")
+        command = command + vlagArg1 + " "
+
+    elif(vlagArg1 == "auto-recovery"):
+        # debugOutput("auto-recovery")
+        command = command + vlagArg1 + " "
+        value = cnos.checkSanityofVariable(
+            deviceType, "vlag_auto_recovery", vlagArg2)
+        if(value == "ok"):
+            command = command + vlagArg2
+        else:
+            retVal = "Error-160"
+            return retVal
+
+    elif(vlagArg1 == "config-consistency"):
+        # debugOutput("config-consistency")
+        command = command + vlagArg1 + " "
+        value = cnos.checkSanityofVariable(
+            deviceType, "vlag_config_consistency", vlagArg2)
+        if(value == "ok"):
+            command = command + vlagArg2
+        else:
+            retVal = "Error-161"
+            return retVal
+
+    elif(vlagArg1 == "isl"):
+        # debugOutput("isl")
+        command = command + vlagArg1 + " port-channel "
+        value = cnos.checkSanityofVariable(
+            deviceType, "vlag_port_aggregation", vlagArg2)
+        if(value == "ok"):
+            command = command + vlagArg2
+        else:
+            retVal = "Error-162"
+            return retVal
+
+    elif(vlagArg1 == "mac-address-table"):
+        # debugOutput("mac-address-table")
+        command = command + vlagArg1 + " refresh"
+
+    elif(vlagArg1 == "peer-gateway"):
+        # debugOutput("peer-gateway")
+        command = command + vlagArg1 + " "
+
+    elif(vlagArg1 == "priority"):
+        # debugOutput("priority")
+        command = command + vlagArg1 + " "
+        value = cnos.checkSanityofVariable(deviceType, "vlag_priority",
+                                           vlagArg2)
+        if(value == "ok"):
+            command = command + vlagArg2
+        else:
+            retVal = "Error-163"
+            return retVal
+
+    elif(vlagArg1 == "startup-delay"):
+        # debugOutput("startup-delay")
+        command = command + vlagArg1 + " "
+        value = cnos.checkSanityofVariable(
+            deviceType, "vlag_startup_delay", vlagArg2)
+        if(value == "ok"):
+            command = command + vlagArg2
+        else:
+            retVal = "Error-164"
+            return retVal
+
+    elif(vlagArg1 == "tier-id"):
+        # debugOutput("tier-id")
+        command = command + vlagArg1 + " "
+        value = cnos.checkSanityofVariable(deviceType, "vlag_tier_id", vlagArg2)
+        if(value == "ok"):
+            command = command + vlagArg2
+        else:
+            retVal = "Error-165"
+            return retVal
+
+    elif(vlagArg1 == "vrrp"):
+        # debugOutput("vrrp")
+        command = command + vlagArg1 + " active"
+
+    elif(vlagArg1 == "instance"):
+        # debugOutput("instance")
+        command = command + vlagArg1 + " "
+        value = cnos.checkSanityofVariable(deviceType, "vlag_instance",
+                                           vlagArg2)
+        if(value == "ok"):
+            command = command + vlagArg2
+            if(vlagArg3 is not None):
+                command = command + " port-channel "
+                value = cnos.checkSanityofVariable(
+                    deviceType, "vlag_port_aggregation", vlagArg3)
+                if(value == "ok"):
+                    command = command + vlagArg3
+                else:
+                    retVal = "Error-162"
+                    return retVal
+            else:
+                command = command + " enable "
+        else:
+            retVal = "Error-166"
+            return retVal
+
+    elif(vlagArg1 == "hlthchk"):
+        # debugOutput("hlthchk")
+        command = command + vlagArg1 + " "
+        value = cnos.checkSanityofVariable(
+            deviceType, "vlag_hlthchk_options", vlagArg2)
+        if(value == "ok"):
+            if(vlagArg2 == "keepalive-attempts"):
+                value = cnos.checkSanityofVariable(
+                    deviceType, "vlag_keepalive_attempts", vlagArg3)
+                if(value == "ok"):
+                    command = command + vlagArg2 + " " + vlagArg3
+                else:
+                    retVal = "Error-167"
+                    return retVal
+            elif(vlagArg2 == "keepalive-interval"):
+                value = cnos.checkSanityofVariable(
+                    deviceType, "vlag_keepalive_interval", vlagArg3)
+                if(value == "ok"):
+                    command = command + vlagArg2 + " " + vlagArg3
+                else:
+                    retVal = "Error-168"
+                    return retVal
+            elif(vlagArg2 == "retry-interval"):
+                value = cnos.checkSanityofVariable(
+                    deviceType, "vlag_retry_interval", vlagArg3)
+                if(value == "ok"):
+                    command = command + vlagArg2 + " " + vlagArg3
+                else:
+                    retVal = "Error-169"
+                    return retVal
+            elif(vlagArg2 == "peer-ip"):
+                # Here I am not taking care of IPV6 option.
+                value = cnos.checkSanityofVariable(
+                    deviceType, "vlag_peerip", vlagArg3)
+                if(value == "ok"):
+                    command = command + vlagArg2 + " " + vlagArg3
+                    if(vlagArg4 is not None):
+                        value = cnos.checkSanityofVariable(
+                            deviceType, "vlag_peerip_vrf", vlagArg4)
+                        if(value == "ok"):
+                            command = command + " vrf " + vlagArg4
+                        else:
+                            retVal = "Error-170"
+                            return retVal
+        else:
+            retVal = "Error-171"
+            return retVal
+
+    else:
+        retVal = "Error-172"
+        return retVal
+
+    # debugOutput(command)
+    cmd = [{'command': command, 'prompt': None, 'answer': None}]
+    retVal = retVal + str(cnos.run_cnos_commands(module, cmd))
+    return retVal
+# EOM
+
+
 def main():
     #
     # Define parameters for vlag creation entry
@@ -290,7 +476,7 @@ def main():
     output = ""
 
     # Send the CLi command
-    output = output + str(cnos.vlagConfig(module, '(config)#', None))
+    output = output + str(vlagConfig(module, '(config)#', None))
 
     # Save it into the file
     file = open(outputfile, "a")
